agent/yozm-ai-agent/pyproject.toml
[project]
name = "yozm-ai-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "load-dotenv>=0.1.0",
    "openai==2.14.0",
    "rich>=14.2.0",
    "ruff>=0.14.10",
    "tenacity>=9.1.2",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]

[tool.ruff.pydocstyle]
convention = "google"
